move big test case to new package use cases where bigger tests will be at
