+++ use_cases/test_singleLineTwoRobots.py
@@ -1,0 +1,123 @@
+from multiprocessing import Manager
+from unittest import TestCase
+
+from swarm_bots.goal.goal_building import GoalBuilding
+from swarm_bots.goal.goal_building_2d import GoalBuilding2D
+from swarm_bots.grid.base_grid import BaseGrid
+from swarm_bots.grid.shared_grid_access import SharedGridAccess
+from swarm_bots.robot_executors.hit_information import HitType
+from swarm_bots.robot_executors.robot_executor import RobotExecutor
+from swarm_bots.robot_executors.spiral.line_scanner_executor import LineScannerExecutor
+from swarm_bots.robot_executors.spiral.line_to_middle import LineToMiddle
+from swarm_bots.tiles.robot import Robot
+from swarm_bots.tiles.tile import Tile, TileType
+from swarm_bots.utils.coordinates import Coordinates
+from swarm_bots.utils.direction import Direction
+
+
+# executor that assumes sth like this in left column of grid
+# source for robotX
+# robotX
+# ...
+# line_to_middle
+# ...
+# robotY
+# source for robotY
+class LineScannerWithSourceWrapperExecutor(RobotExecutor):
+    def __init__(self, line: LineToMiddle, robot: Robot, shared_grid_access: SharedGridAccess,
+                 goal_building: GoalBuilding):
+        super().__init__(robot, shared_grid_access, goal_building, 0.0001)
+        self.line = line.copy()
+        self.line_scanner = LineScannerExecutor(self.shared_actions_executor)
+
+    def start_process(self):
+        start_coordinates = self.robot_coordinates.copy()
+        to_line_start_direction = start_coordinates.get_to_other_direction(self.line.start_coordinates)
+        to_source_direction = to_line_start_direction.get_opposite()
+        while not self.line.is_finished():
+            self.shared_actions_executor.wait_action()
+            self.shared_actions_executor.wait_action()
+            self.shared_actions_executor.wait_action()
+            if not self.robot.rotation == to_source_direction:
+                self.shared_actions_executor.try_rotate_robot(to_source_direction)
+            if not self.robot.has_block():
+                self.shared_actions_executor.try_get_block(to_source_direction)
+            if not self.robot.has_block():
+                raise ValueError("robot could not take block")
+            self.shared_actions_executor.try_rotate_robot(to_line_start_direction)
+            while self.robot_coordinates != self.line.start_coordinates:
+                hit_information = self.shared_actions_executor.try_move_robot(to_line_start_direction)
+                if hit_information.hit_type == HitType.ROBOT:
+                    self.shared_actions_executor.wait_action()
+            self.line_scanner.scan_line(self.line)
+            self.shared_actions_executor.try_rotate_robot(to_line_start_direction.get_opposite())
+            while self.robot_coordinates != start_coordinates:
+                hit_information = self.shared_actions_executor.try_move_robot(to_source_direction)
+                if hit_information.hit_type == HitType.ROBOT:
+                    self.shared_actions_executor.wait_action()
+        print(f"robot: {self.robot} put {self.shared_actions_executor.put_blocks} blocks")
+
+
+class TestSingleLineTwoRobots(TestCase):
+    # noinspection PyPep8
+    def test_robot_crash_updated_tiles_not_by_self_with_sleeps(self):
+        text_grid = """
+        0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
+        0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
+        0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
+        0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
+        0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
+        0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
+        0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
+        0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
+        0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 1 0 0 0 1 0 0 0 0 0 1 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 1 1 0 0 1 1 1 0 1 0 0 0 0 0 0 1 1 1 0 0 0 0 0 0 0 0 0 0 0 1 1 1 1 1 0 0 0 0 0 0 0 0 0 0 0 0 0 1
+        0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
+        0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
+        0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
+        0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
+        0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
+        0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
+        0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
+        0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
+        0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
+        """
+        goal_building = GoalBuilding2D(text_grid=text_grid)
+        robot_1 = Robot(Direction.DOWN)
+        robot_1_source = Tile(TileType.SOURCE)
+        robot_2 = Robot(Direction.UP)
+        robot_2_source = Tile(TileType.SOURCE)
+        line_start_coordinates = Coordinates(0, 9)
+        base_grid = BaseGrid(goal_building.width, goal_building.height)
+        robot_1_coordinates = Coordinates(0, 16)
+        base_grid.add_tile_to_grid(robot_1, robot_1_coordinates)
+        base_grid.add_tile_to_grid(robot_1_source, robot_1_coordinates.create_neighbour_coordinate(Direction.UP))
+        robot_2_coordinates = Coordinates(0, 1)
+        base_grid.add_tile_to_grid(robot_2, robot_2_coordinates)
+        base_grid.add_tile_to_grid(robot_2_source, robot_2_coordinates.create_neighbour_coordinate(Direction.DOWN))
+
+        shared_grid_access = SharedGridAccess(base_grid, manager=Manager())
+        line_to_middle = LineToMiddle(
+            start_coordinates=line_start_coordinates,
+            direction=Direction.RIGHT,
+            block_line=list(map(bool, map(int, text_grid.split("\n")[9].split())))
+        )
+        robot_1_executor = LineScannerWithSourceWrapperExecutor(
+            line=line_to_middle,
+            robot=robot_1,
+            shared_grid_access=shared_grid_access,
+            goal_building=goal_building
+        )
+        robot_2_executor = LineScannerWithSourceWrapperExecutor(
+            line=line_to_middle,
+            robot=robot_2,
+            shared_grid_access=shared_grid_access,
+            goal_building=goal_building
+        )
+        robot_1_executor.start_working()
+        robot_2_executor.start_working()
+
+        robot_1_executor.wait_for_finish()
+        robot_2_executor.wait_for_finish()
+
+        grid = shared_grid_access.get_private_copy()
+        assert goal_building.validate_grid(grid)
